# [m3msg] Improve message scan performance